Add files via upload
Financial institutions and payment operators are increasingly relying on machine learning algorithms to develop efficient and effective fraud detection systems. The main aim is to implement and assess the performance of various machine learning models, including logistic regression, random forests, and neural networks, etc. A key challenge with payment fraud data is a class imbalance. Class imbalance can make it difficult for standard models to learn to distinguish between the majority and minority classes. Due to privacy concerns, it is not possible to infer meaningful relationships between most of the variables in the dataset. The main aim of the project would be to come with the best predictive model rather than an inference model.